Bundle user && account page
fix username refresh onChange and admin privileges
