+++ src/components/Modal/modalTypes.js
@@ -6,3 +6,5 @@ export const PAYWALL_MODAL = "PAYWALL_MODAL";
 export const PURCHASE_CREDITS_MODAL = "PURCHASE_CREDITS_MODAL";
 export const CREDITS_HISTORY_MODAL = "CREDITS_HISTORY_MODAL";
 export const MANAGE_CREDITS_MODAL = "MANAGE_CREDITS_MODAL";
+export const CHANGE_USERNAME_MODAL = "CHANGE_USERNAME_MODAL";
+export const CHANGE_PASSWORD_MODAL = "CHANGE_PASSWORD_MODAL";
