+++ src/components/UserDetail/index.js
@@ -2,14 +2,14 @@ import React, { Component } from "react";
 import { autobind } from "core-decorators";
 import UserDetailPage from "./Page";
 import userConnector from "../../connectors/user";
-import { USER_DETAIL_TAB_PROPOSALS } from "../../constants";
+import { USER_DETAIL_TAB_PROPOSALS, USER_DETAIL_TAB_GENERAL } from "../../constants";
 
 class UserDetail extends Component {
 
   constructor(props) {
     super(props);
     this.state = {
-      tabId: USER_DETAIL_TAB_PROPOSALS
+      tabId: USER_DETAIL_TAB_GENERAL
     };
   }
 
@@ -18,15 +18,24 @@ class UserDetail extends Component {
     this.props.onFetchProposalsVoteStatus();
   }
 
-  componentDidUpdate() {
-    const { editUserResponse } = this.props;
+  componentDidUpdate(prevProps) {
+    const { loggedInAsUserId, editUserResponse, user, isAdmin } = this.props;
     if(editUserResponse) {
       window.location.reload();
+    }
+
+    const isAdminOrTheUser = user && (isAdmin || loggedInAsUserId === user.id);
+    if (!isAdminOrTheUser && (prevProps.user !== this.props.user)) {
+      this.setState({ tabId: USER_DETAIL_TAB_PROPOSALS });
     }
   }
 
   onTabChange(tabId) {
     this.setState({ tabId: tabId });
+  }
+
+  componentWillUnmount() {
+    this.unmounting = true;
   }
 
   render() {
