+++ src/components/snew/SubmitPage.js
@@ -50,6 +50,7 @@ class SubmitPage extends React.Component {
       newProposalError,
       userCanExecuteActions,
       openModal,
+      userid,
       proposalCredits,
       editingMode
     } = this.props;
@@ -59,7 +60,7 @@ class SubmitPage extends React.Component {
         <div className="page submit-proposal-page">
           <ReactBody className="submit-page" />
           <div
-            className="submit content warn-on-unload"
+            className="submit conztent warn-on-unload"
             id="newlink"
           >
             {newProposalError && (
@@ -148,6 +149,9 @@ class SubmitPage extends React.Component {
                           To submit a proposal, you must purchase a proposal credit.
                           <a className="linkish" onClick={() => openModal(MANAGE_CREDITS_MODAL)}> Click here</a>{" "}
                           to open the proposal credits manager.
+                          To submit a proposal, you must purchase a proposal credit,
+                          see your <a href={`/user/${userid}`}>account page</a>{" "}
+                          for more information.
                         </div>
                       )}
                     </div>
