--- src/connectors/account.js
@@ -1,34 +1,0 @@
-import { connect } from "react-redux";
-import * as sel from "../selectors";
-import * as act from "../actions";
-import { withRouter } from "react-router-dom";
-import { onUpdateUserKey } from "../actions/api";
-import { onIdentityImported } from "../actions/app";
-import { bindActionCreators } from "redux";
-import { confirmWithModal } from "../actions/modal";
-import compose from "lodash/fp/compose";
-
-const accountConnector = connect(sel.selectorMap({
-  loggedInAsEmail: sel.loggedInAsEmail,
-  userAlreadyPaid: sel.userAlreadyPaid,
-  paywallAddress: sel.paywallAddress,
-  paywallAmount: sel.paywallAmount,
-  paywallTxNotBefore: sel.paywallTxNotBefore,
-  keyMismatch: sel.getKeyMismatch,
-  updateUserKey: sel.updateUserKey,
-  updateUserKeyError: sel.updateUserKeyError,
-  verificationToken: sel.verificationToken,
-  shouldAutoVerifyKey: sel.shouldAutoVerifyKey,
-  identityImportError: sel.identityImportError,
-  identityImportSuccess: sel.identityImportSuccess,
-  userPubkey: sel.userPubkey,
-  userID: sel.userid
-}), dispatch => bindActionCreators({
-  onUpdateUserKey,
-  keyMismatchAction: act.keyMismatch,
-  onIdentityImported,
-  confirmWithModal
-}, dispatch)
-);
-
-export default compose(withRouter, accountConnector);
